// redditclone-frontend/client-codegen-api/models/register-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reddit Clone API
 * Clone Reddit With Spring Boot.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface RegisterRequestDto
 */
export interface RegisterRequestDto {
    /**
     * 
     * @type {string}
     * @memberof RegisterRequestDto
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof RegisterRequestDto
     */
    username?: string;
    /**
     * 
     * @type {string}
     * @memberof RegisterRequestDto
     */
    password?: string;
}